abstract logic to vault filter service
