+++ apps/web/src/app/vault/individual-vault/vault-filter/vault-filter.module.ts
@@ -1,5 +1,6 @@
 import { NgModule } from "@angular/core";
 
+import { VaultFilterService as SharedVaultFilterService } from "@bitwarden/angular/vault/vault-filter/services/vault-filter.service";
 import { SearchModule } from "@bitwarden/components";
 
 import { VaultFilterSharedModule } from "../../individual-vault/vault-filter/shared/vault-filter-shared.module";
@@ -17,6 +18,10 @@ import { VaultFilterService } from "./services/vault-filter.service";
     {
       provide: VaultFilterServiceAbstraction,
       useClass: VaultFilterService,
+    },
+    {
+      provide: SharedVaultFilterService,
+      useClass: SharedVaultFilterService,
     },
   ],
 })
